Move build_option and build_ok_or_else to emit::func
